CrewAI/CrewAI_DataC/pyproject.toml
[project]
name = "crewai-datacamp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.10.1",
    "crewai-tools>=1.10.1",
    "groq>=1.0.0",
    "ipykernel>=7.2.0",
    "litellm>=1.82.0",
    "python-dotenv>=1.1.1",
]
